[frontend] Improve Edition user experience (#5844)
